organizes scrape code into sub dir
